src/agents/orchestrator-agent/pyproject.toml
[project]
name = "orchestrator-agent"
version = "0.1.0"
description = "Orchestrator Agent A2A"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    # A2A Protocol
    "a2a-sdk>=0.3.7",
    "a2a-sdk[http-server]>=0.3.7",
    
    "python-dotenv>=1.0.0",
    "qdrant-client>=1.7.0",
    "sentence-transformers>=2.2.0",
    "torch>=2.0.0",
    "transformers>=4.30.0",

    # LangChain
    "langchain>=0.1.0",
    "langchain-huggingface>=0.2.0",
    "langchain-google-genai==2.1.9",
    "langchain-core>=0.1.0",
    "langgraph>=0.6.5",

    "uvicorn>=0.35.0",
    "pydantic>=2.11.5",
    "fastapi>=0.118.0",

    "redis>=6.3.0"
]

[tool.uv.workspace]
members = [
    "orchestrator-agent",
]
